+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * [HexaPDF::Task::ImportPages] for importing pages while preserving the optional
   content configuration and interactive form fields
 * [HexaPDF::Type::FontType3#embedded?] as Type 3 fonts are always embedded
+* CLI options `--acro-form` and `--optional-content` to `hexapdf merge` for
+  specifying how interactive form fields and optional content are handled
 
 ### Changed